requirements.txt
youtube-transcript-api
google-api-python-client
google-generativeai
python-dotenv
